ln-simln-jamming: bootstrap attacker for custom period
To allow simulations that have an attacker that's been around for less
time than the rest of the network, we separate the bootstrap of attacker
revenue from bootstrap of the rest of the network. This can be set to
the same period by just setting the attacker's bootstrap period to the
revenue window used by the simulation.