ln-simln-jamming: use peacetime traffic file if not bootstrapping
If there is no need for the attacker to sit and build reputation in
the network for an attack, just use the peacetime traffic file to
bootstrap the reputation of honest nodes.

This allows a quicker start for this type of attack, because a separate
traffic file that has traffic including the attacker does not need to
be generated.