ln-resource-mgr: track utilization of protected/congested slots/capacity
We're interested in knowing how much our protected resources are used,
because it helps us understand what the opportunity cost of allowing
htlcs access to them is.

This isn't easily abusable because you have to have reputation to get
into these buckets in the first place (/have very limited access to
congestion) and have to pay fees to impact the value.